Optimise markdown subsystem memory + allocation
Cut the markdown data structures' footprint and per-frame allocation churn
without changing behaviour:

- Pack booleans as C++20 bitfields (member-access syntax + defaulted == are
  unchanged): TextStyle, CodeRun, PaintLine, PaintRunState, LaidScrollBlock,
  MdBlock.
- Reorder PaintLine / LaidScrollBlock / MdBlock largest-alignment-first to drop
  interior padding; shrink MdBlock::headingLevel to uint8_t and headerRowCount
  to uint32_t.
- De-duplicate held text: MarkdownView no longer keeps its own copy of the
  source (delegates to MarkdownDocument); LaidScrollBlock no longer stores the
  full per-fence codeText — it keeps a codeLineCount + codeContentHash computed
  once at layout (the body is re-resolved on demand via resolveCodeBlockText).
- Cache the palette + selection-highlight colour in MarkdownView (were rebuilt
  on every onPaint) — refreshed in resolveFonts on theme change.

Builds with -Werror; 615/615 non-GUI tests pass.

+++ src/lib/markdown/Markdown.cpp
@@ -259,7 +259,7 @@ auto mdEnterBlock(const MD_BLOCKTYPE type, void* detail, void* userData) -> int 
     case MD_BLOCK_H: {
         const auto* hDetail = static_cast<MD_BLOCK_H_DETAIL*>(detail);
         builder.begin(MdBlockKind::Heading);
-        builder.cur.headingLevel = hDetail->level;
+        builder.cur.headingLevel = static_cast<std::uint8_t>(hDetail->level);
         break;
     }
     case MD_BLOCK_CODE: {
